requirements.txt
Flask=3.0.3
Pillow
pytesseract



# Set-ExecutionPolicy -ExecutionPolicy RemoteSigned -Scope Process
#  .\venv\Scripts\Activate.ps1  